Refactor make business logic out of AdminCategoriesController
