+++ SunEngine/Admin/Controllers/AdminCategoriesController.cs
@@ -19,19 +19,16 @@ namespace SunEngine.Admin.Controllers
 {
     public class AdminCategoriesController : AdminBaseController
     {
-        private readonly DataBaseConnection db;
         private readonly ICategoriesCache categoriesCache;
         private readonly CategoriesManager categoriesManager;
         private readonly ICategoriesAdminPresenter categoriesAdminPresenter;
 
         public AdminCategoriesController(
-            DataBaseConnection db,
             CategoriesManager categoriesManager,
             ICategoriesAdminPresenter categoriesAdminPresenter,
             ICategoriesCache categoriesCache,
             MyUserManager userManager, IRolesCache rolesCache) : base(rolesCache, userManager)
         {
-            this.db = db;
             this.categoriesCache = categoriesCache;
             this.categoriesManager = categoriesManager;
             this.categoriesAdminPresenter = categoriesAdminPresenter;
@@ -61,38 +58,7 @@ namespace SunEngine.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> GetAllCategories()
         {
-            var categories = await db.Categories.OrderBy(x => x.SortNumber).Select(x => new CategoryAdminViewModel
-            {
-                Id = x.Id,
-                Name = x.Name,
-                Title = x.Title,
-                IsFolder = x.IsFolder,
-                IsMaterialsContainer = x.IsMaterialsContainer,
-                Description = x.Description,
-                Header = x.Header,
-                AreaRoot = x.AreaRoot,
-                ParentId = x.ParentId,
-                SortNumber = x.SortNumber,
-                MaterialsCount = x.Materials.Count,
-                IsHidden = x.IsHidden,
-                IsDeleted = x.IsDeleted
-            }).ToDictionaryAsync(x => x.Id);
-
-            CategoryAdminViewModel root = null;
-
-            foreach (var category in categories.Values)
-            {
-                if (!category.ParentId.HasValue)
-                {
-                    root = category;
-                    continue;
-                }
-
-                var parent = categories[category.ParentId.Value];
-                if (parent.SubCategories == null) parent.SubCategories = new List<CategoryAdminViewModel>();
-
-                parent.SubCategories.Add(category);
-            }
+            var root = await categoriesAdminPresenter.GetAllCategoriesAsync();
 
             return Json(root);
         }
@@ -130,59 +96,25 @@ namespace SunEngine.Admin.Controllers
         [HttpPost]
         public async Task<IActionResult> CategoryUp(string name)
         {
-            using (var transaction = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
-            {
-                var category = await db.Categories.FirstOrDefaultAsync(x => x.Name == name);
-                if (category == null)
-                    return BadRequest();
+            var rez = await categoriesManager.CategoryUp(name);
+            if (rez.Failed)
+                return BadRequest();
 
-                var category2 = await db.Categories
-                    .Where(x => x.ParentId == category.ParentId && x.SortNumber < category.SortNumber)
-                    .OrderByDescending(x => x.SortNumber).FirstOrDefaultAsync();
+            categoriesCache.Reset();
 
-                if (category2 == null)
-                    return BadRequest();
-
-                await db.Categories.Where(x => x.Id == category.Id).Set(x => x.SortNumber, category2.SortNumber)
-                    .UpdateAsync();
-                await db.Categories.Where(x => x.Id == category2.Id).Set(x => x.SortNumber, category.SortNumber)
-                    .UpdateAsync();
-
-                transaction.Complete();
-
-                categoriesCache.Reset();
-
-                return Ok();
-            }
+            return Ok();
         }
 
         [HttpPost]
         public async Task<IActionResult> CategoryDown(string name)
         {
-            using (var transaction = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
-            {
-                var category = await db.Categories.FirstOrDefaultAsync(x => x.Name == name);
-                if (category == null)
-                    return BadRequest();
+            var rez = await categoriesManager.CategoryDown(name);
+            if (rez.Failed)
+                return BadRequest();
 
-                var category2 = await db.Categories
-                    .Where(x => x.ParentId == category.ParentId && x.SortNumber > category.SortNumber)
-                    .OrderBy(x => x.SortNumber).FirstOrDefaultAsync();
+            categoriesCache.Reset();
 
-                if (category2 == null)
-                    return BadRequest();
-
-                await db.Categories.Where(x => x.Id == category.Id).Set(x => x.SortNumber, category2.SortNumber)
-                    .UpdateAsync();
-                await db.Categories.Where(x => x.Id == category2.Id).Set(x => x.SortNumber, category.SortNumber)
-                    .UpdateAsync();
-
-                transaction.Complete();
-
-                categoriesCache.Reset();
-
-                return Ok();
-            }
+            return Ok();
         }
 
         [HttpPost]
@@ -191,36 +123,5 @@ namespace SunEngine.Admin.Controllers
             categoriesCache.Reset();
             return Ok();
         }
-    }
-
-    public class CategoryAdminViewModel
-    {
-        public int Id { get; set; }
-
-        public string Name { get; set; }
-
-        public string Title { get; set; }
-
-        public bool IsFolder { get; set; }
-
-        public bool IsMaterialsContainer { get; set; }
-
-        public string Description { get; set; }
-
-        public string Header { get; set; }
-
-        public bool AreaRoot { get; set; }
-
-        public int? ParentId { get; set; }
-
-        public IList<CategoryAdminViewModel> SubCategories { get; set; }
-
-        public int SortNumber { get; set; }
-
-        public int MaterialsCount { get; set; }
-
-        public bool IsDeleted { get; set; }
-
-        public bool IsHidden { get; set; }
     }
 }
