+++ SunEngine/Admin/Managers/CategoriesManager.cs
@@ -6,6 +6,7 @@ using LinqToDB;
 using SunEngine.DataBase;
 using SunEngine.Models;
 using SunEngine.Services;
+using SunEngine.Stores;
 using SunEngine.Utils.TextProcess;
 
 namespace SunEngine.Admin.Managers
@@ -13,12 +14,16 @@ namespace SunEngine.Admin.Managers
     public class CategoriesManager : DbService
     {
         private readonly Sanitizer sanitizer;
+        private readonly ICategoriesCache categoriesCache;
+
         
         public CategoriesManager(
             DataBaseConnection db,
+            ICategoriesCache categoriesCache,
             Sanitizer sanitizer) : base(db)
         {
             this.sanitizer = sanitizer;
+            this.categoriesCache = categoriesCache;
         }
 
         public async Task AddCategoryAsync(Category category)
@@ -57,6 +62,58 @@ namespace SunEngine.Admin.Managers
             сategoryExisted.AreaRoot = category.AreaRoot;
             
             await db.UpdateAsync(сategoryExisted);
+        }
+        
+        public async Task<ServiceResult> CategoryUp(string name)
+        {
+            using (var transaction = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
+            {
+                var category = await db.Categories.FirstOrDefaultAsync(x => x.Name == name);
+                if (category == null)
+                    return ServiceResult.BadResult();
+
+                var category2 = await db.Categories
+                    .Where(x => x.ParentId == category.ParentId && x.SortNumber < category.SortNumber)
+                    .OrderByDescending(x => x.SortNumber).FirstOrDefaultAsync();
+
+                if (category2 == null)
+                    return ServiceResult.BadResult();
+
+                await db.Categories.Where(x => x.Id == category.Id).Set(x => x.SortNumber, category2.SortNumber)
+                    .UpdateAsync();
+                await db.Categories.Where(x => x.Id == category2.Id).Set(x => x.SortNumber, category.SortNumber)
+                    .UpdateAsync();
+
+                transaction.Complete();
+            }
+
+            return ServiceResult.OkResult();
+        }
+
+        public async Task<ServiceResult> CategoryDown(string name)
+        {
+            using (var transaction = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
+            {
+                var category = await db.Categories.FirstOrDefaultAsync(x => x.Name == name);
+                if (category == null)
+                    return ServiceResult.BadResult();
+
+                var category2 = await db.Categories
+                    .Where(x => x.ParentId == category.ParentId && x.SortNumber > category.SortNumber)
+                    .OrderBy(x => x.SortNumber).FirstOrDefaultAsync();
+
+                if (category2 == null)
+                    return ServiceResult.BadResult();
+
+                await db.Categories.Where(x => x.Id == category.Id).Set(x => x.SortNumber, category2.SortNumber)
+                    .UpdateAsync();
+                await db.Categories.Where(x => x.Id == category2.Id).Set(x => x.SortNumber, category.SortNumber)
+                    .UpdateAsync();
+
+                transaction.Complete();
+            }
+            
+            return ServiceResult.OkResult();
         }
     }
 }
